refactor(server): share grantPermission test helper across route tests.
